@@ -1,29 +1,41 @@
 apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
+apply plugin: 'org.jetbrains.kotlin.android'
+apply plugin: 'org.jetbrains.kotlin.plugin.compose'
 apply plugin: 'de.undercouch.download'
 apply plugin: 'maven-publish'
 
+
 android {
-    compileSdkVersion 34
+    namespace "com.assentify.sdk"
+    compileSdkVersion 36
+    android.ndkVersion "27.0.12077973"
     defaultConfig {
-      applicationId "com.assentify.sdk"
+         applicationId "com.assentify.sdk"
 
         minSdkVersion 24
         targetSdkVersion 34
         versionCode 1
         versionName "1.0"
         multiDexEnabled = true
-        android.ndkVersion "25.1.8937393"
+
     }
 
 
-
+/*    signingConfigs {
+        create("release") {
+            keyAlias = "release-key"
+            keyPassword = "00000000"
+            storeFile = file("release-key.jks")
+            storePassword = "00000000"
+        }
+    }*/
 
     buildTypes {
         release {
-            minifyEnabled true
-            shrinkResources = true
+            minifyEnabled false
+            // shrinkResources = true
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+            //signingConfig = signingConfigs.getByName("release")
         }
     }
 
